Fix child classpath of MapReduce child task jvms.
Maven uses a ridiculous classloader to load jars in at runtime.
This completely breaks the in-process MapReduce cluster because
it determines the classpath for the child jvms of TaskAttempts
assuming the normal classloader.

To get around this, we set the classpath programatically to have
both the dependencies of the plugin and the maven project's test
classpath.