hbase-maven-plugin writes seperate core-site.xml file.
Previously, the hbase-maven-plugin, after startup, would write an hbase-site.xml file containing
all configuration used to start the HDFS, MapReduce, and HBase clusters to the test-classes
folder of the build directory of the project. There's a disadvantage to this, in that client test
code must add HBase resources to a Configuratin object in order to get HDFS and MapReduce
configuration, even if they don't intend to use HBase.

This commit changes the hbase-maven-plugin so that it writes both an hbase-site.xml and
a core-site.xml file.  core-site.xml contains MapReduce/HDFS configuration, and hbase-site.xml
contains HBase-specific configuration. This has been tested by using the plugin in a sample
project and verifying that creating a new configuration adds the configuration found in
core-site.xml.