Adds modified FSUtils.java to suppress superfluous stack traces.
HBase 0.92.1, included with CDH4.1.2, contains a harmless race condition when
deleting tables, resulting in stack traces hardcoded to print to standard out.

This change includes a custom FSUtils which LOG.warns instead.

This change is specific to HBase 0.92.1. It should not be cherry picked on
other branches thoughtlessly and it should be reverted if the hbase dependency
upgrades to 0.92.2.

Signed-off-by: Aaron Kimball <aaron@wibidata.com>